removed classic redux pattern and used redoxtoolkit for userSlice
