+++ src/app/rootReducer.js
@@ -1,8 +1,8 @@
 import { combineReducers } from '@reduxjs/toolkit'
-import authSlice from '../auth/authSlice';
+import userSlice from '../auth/userSlice';
 
 const rootReducer = combineReducers({
-  auth: authSlice,
+  user: userSlice,
 });
 
 export default rootReducer;
