--- src/redux/rootReducer.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import userReducer from './user/user.reducer';
-
-export default combineReducers({
-  user: userReducer,
-});
